refactor(cli): extract WebUI support
